Refactor Left and Right Tuples aligned with Super Cache improvements #5648 (#5649)
* -BetaConstraints no longer depends on ReteEvaluator.

* -Trying to separate BetaMemory so BetaConstraints can go in -base.

* -Refactoring Tuples around concrete classes.

* -Centralise and improve super cache handling fix.
-Improve NodeTypeEnums to contain more information (at no cost) and replace as many instanceof as possible.

* -Move Super Cache helper methods into standalone class SuperCacheFixer.

* -Remove RightTuple interface.

* -Refactory creation to a Factory with switch for different implementations.

* -Added missing headers.
-Remove unecessary if statement with printout.
-Removed incorrectly added maven module.

* minor semplifications

---------

+++ drools-base/src/main/java/org/drools/base/rule/IndexableConstraint.java
@@ -30,7 +30,7 @@ public interface IndexableConstraint extends Constraint {
 
     boolean isUnification();
 
-    boolean isIndexable(short nodeType, KieBaseConfiguration config);
+    boolean isIndexable(int nodeType, KieBaseConfiguration config);
 
     ConstraintTypeOperator getConstraintType();
